refactor(utils): :recycle: refactor utils
